backend/requirements.txt
openai 1.40.6
python-decouple 3.8
python-multipart 0.0.19
requests 2.32.3
fastapi 0.115.5
"uvicorn[standard]"
python-dotenv 1.01
langchain 0.3.9
langchain_openai 0.2.10
langchain-core 0.3.21
langchain_text_splitters 0.3.2
langchain_community 0.3.9
faiss-cpu 1.9.0
nltk 3.9.1
inflect 7.4.0
rank_bm25 0.2.2

